Move CalculationStatusEnum, CalculationStatusDetails into different module
