+++ pyqcrbox/pyqcrbox/sql_models_NEW_v2/__init__.py
@@ -2,6 +2,7 @@ from .application_spec import ApplicationSpec, ApplicationSpecWithCommands
 from .application_spec_db import ApplicationSpecDB
 from .base import QCrBoxBaseSQLModel, QCrBoxPydanticBaseModel
 from .calculation import CalculationDB
+from .calculation_status_event import CalculationStatusEnum, CalculationStatusDetails
 from .command_execution import CommandExecutionCreate
 from .command_invocation import CommandInvocationCreate
 from .command_spec import CommandSpec, CommandSpecDB, CommandSpecWithParameters
